requirements.txt
streamlit>=1.20
pdfplumber>=0.7.5
pandas>=1.5
scikit-learn>=1.2
numpy>=1.23
openpyxl>=3.1.0
python-dotenv>=1.0.0
